+++ apps/pawfinder/lib/pages/welcome_page.dart
@@ -1,16 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:services/providers/auth_provider.dart';
 import 'package:material_symbols_icons/material_symbols_icons.dart';
+import 'package:services/providers/auth_provider.dart';
 
 class WelcomePage extends ConsumerWidget {
   const WelcomePage({super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final auth = ref.read(authProvider);
-
     return Scaffold(
       body: Center(
         child: Column(
@@ -37,6 +35,8 @@ class WelcomePage extends ConsumerWidget {
                   FilledButton.icon(
                         onPressed: () async {
                           try {
+                            final auth = await ref.read(authProvider.future);
+
                             await auth.login([
                               'openid',
                               'profile',
@@ -47,7 +47,7 @@ class WelcomePage extends ConsumerWidget {
                           } on OfflineAuthException {
                             if (context.mounted) {
                               ScaffoldMessenger.of(context).showSnackBar(
-                                SnackBar(
+                                const SnackBar(
                                   content: Text('No internet connection'),
                                 ),
                               );